Move sync_user_case stuff to separate file
